ref: cleanup + typehints
