+++ elasticai/explorer/cost_estimator.py
@@ -1,23 +1,18 @@
 import torch
 from nni.nas.profiler.pytorch.flops import FlopsProfiler
+from nni.nas.nn.pytorch import ModelSpace
+from nni.mutable.symbol import SymbolicExpression
 
 
 class FlopsEstimator:
-    """Wrapper for FlopsProfiler could extend in future"""
+    """Wrapper for FlopsProfiler"""
 
-    def estimate_flops(self, model_sample: torch.nn.Module) -> int:
-        """Computes FLOPS for a single module.
+    def estimate_flops(self, model_space: ModelSpace) -> SymbolicExpression | float:
+        """Computes FLOPS for a frozen model space."""
 
-        Args:
-            model_sample (torch.nn.Module): A frozen sample from ModelSpace
-
-        Returns:
-            int: The FLOPS-estimate
-        """
-
-        first_parameter = next(model_sample.parameters())
+        first_parameter = next(model_space.parameters())
         input_shape = first_parameter.size()
         data_sample = torch.full(input_shape, 1.0)
-        profiler = FlopsProfiler(model_sample, data_sample)
+        profiler = FlopsProfiler(model_space, data_sample)
 
         return profiler.expression
