+++ elasticai/explorer/platforms/generator/generator.py
@@ -2,13 +2,14 @@ import logging
 import os
 from abc import ABC, abstractmethod
 from pathlib import Path
+from typing import Any
 import torch
 from torch import nn
 
 
 class Generator(ABC):
     @abstractmethod
-    def generate(self, model: nn.Module, path: Path) -> any:
+    def generate(self, model: nn.Module, path: Path) -> Any:
         pass
